refactor(gitlab): break the setup/verifier import cycle via a client leaf
