+++ integrations/gitlab/verifier.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from integrations.gitlab import build_gitlab_config, validate_gitlab_config
+from integrations.gitlab.client import build_gitlab_config, validate_gitlab_config
 from integrations.verification import register_validation_verifier
 
 verify_gitlab = register_validation_verifier(
